Reorganized yices1 target.
Now the yices1 target deals with monomorphizing and sorting and inlining and
simplification and such.

Note: I removed code which was previously used to hack around a bug in yices1
with data declarations across pushes and pops. We may need to add something
like that back in if we run into that bug again.