Changed LaceEH to be LamEH and CaseEH.
That is, in the elaborator we now use desugared case expressions. The hope is
this will eliminate the need for pushing constructors, so that we can
elaborate down to SAT without doing an exhaustive search of the assignments.

Currently only the Basic tests work. I'll have to add some argument and
function pushing before the SMT tests work. That's the next step.