Don't desugar clauses into uses of curry.
Because it was making it so every function we call we would first wrap all the
arguments in a tuple, then deconstruct the tuple to get at the arguments. It's
anticipated this has improved performance a lot. I saw initial evidence of
that in Sudoku. Either way, it's a cleaner implementation anyway, so this is a
good change.